requirements.txt
absl-py==1.4.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==23.1.0
auditwheel==5.3.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==1.3.0
colorama==0.4.6
comm==0.1.3
ConfigArgParse==1.5.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
einops==0.6.1
entrypoints==0.3
exceptiongroup==1.1.1
executing==1.2.0
Farama-Notifications==0.0.4
fastjsonschema==2.16.3
filelock==3.12.0
fonttools==4.39.3
fsspec==2023.5.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
grpcio==1.54.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonschema==4.17.3
kaleido==0.1.0
kiwisolver==1.4.4
lit==16.0.5.post0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook-shim==0.2.3
novograd==0.0.6
numpy==1.24.4
oauthlib==3.2.2
open3d==0.13.0
opencv-python==4.8.0.76
overrides==7.3.1
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
patchelf==0.17.2.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.5.0
plotly==5.14.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pyelftools==0.29
pygame==2.1.3
pyglet==1.5.0
Pygments==2.15.1
pyquaternion==0.9.9
pyrsistent==0.19.3
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pyxdg==0.26
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.2.2
QtPy==2.0.0
requests==2.30.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rsa==4.9
Rtree==0.9.4
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.21.1
setproctitle==1.3.2
sip==4.19.21
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sumolib==1.17.0
sympy==1.11.1
tenacity==8.2.2
termcolor==2.3.0
terminado==0.17.1
testpath==0.5.0
threadpoolctl==3.1.0
timm==0.3.2
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.8
torch==2.0.1
torchaudio==0.13.1+cu116
torchvision==0.15.2
tornado==6.3.1
tqdm==4.65.0
traci==1.17.0
traitlets==5.9.0
triton==2.0.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wandb==0.15.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.3.3
widgetsnbextension==4.0.7
y-py==0.5.9
yacs==0.1.8
ypy-websocket==0.8.4
zipp==3.15.0
